test(server): delete field not working
